Major Mods
I have (and continue) to make some major changes to SSK in preparation
for some new products I will be releasing today and afterwards.

What does this mean to you?  If you have been using SSK and update to
this version, a few function calls may fail.

To fix this, open your code, then go to the SSKCorona Wiki and see if
the function signature has changed. Update your usage to match the new
signature.
https://github.com/roaminggamer/SSKCorona/wiki

Additionally, your custom presets may fail to load.  If they do, compare
the top of your presets file to the ones found in 'SSKCorona\3-Game
Frame\data\*'
You will notice that the 'mgr' loading lines have changed.  Just update
that line and you should be golden.

Thanks and sorry for any issues this may cause.  I promise this change
is for the best.